to change a flow of  individual component according to document and working on Right side menu Pgn List component
