+++ imports/ui/pages/Leftsidebar/PlayComponent.js
@@ -1,0 +1,1 @@
+import Modal from 'react-bootstrap';
